Use PIPE_READ_END, PIPE_WRITE_END to clarify use of privileged op sockets
Both sockets in this socket pair are technically bidirectional, but
we're using them in a way that is close enough to unidirectional that
using these symbolic constants is clearer than magic numbers.
We send multi-byte requests into the write end, and read those requests
from the read end (even though we also send a 1-byte reply to each request
into the "read" end, and read it from the "write" end).

Signed-off-by: Simon McVittie <smcv@collabora.com>