Support scoping abstract unix sockets
It's desirable in many cases to be able to allow a sandboxed program to
exist with the current network namespace without permitting it access
to all abstract unix sockets in said namespace. For example, X11 has an
abstract unix socket @/tmp/.X11-unix/X0, which, using the abs scoping
options this patch introduces, would be inaccessible to a sandboxed
client that resides in the same network namespace as the X11 server.

As we are relied on by various higher level sandboxing frameworks, such
as Glycin and Flatpak, also introduce a `-try` variant that does not
simply bail if unable to restrict access to said unix sockets.

Closes: https://github.com/containers/bubblewrap/issues/330
Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>