Drop check for setcaps
Before bwrap was installed setuid (when user namespaces are not
supported), it used setcap/file capabilities to acquire the privileges
to create the bwrap environment. This was dropped 9 years ago in
aedd613 ("Completely drop setcaps codepaths in favour of setuid").
and the check was added instead.

Given that it has been 9 years, it is exceedingly unlikely that someone
still has misconfigured bwrap in that particular way, so remove the
check.

Signed-off-by: Sebastian Wick <sebastian.wick@redhat.com>